move resample to functional, compare to librosa
